+++ src/main/java/com/ddobang/backend/domain/theme/dto/response/ThemeDetailResponse.java
@@ -1,30 +1,31 @@
-package com.ddobang.backend.domain.theme.dto;
+package com.ddobang.backend.domain.theme.dto.response;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
+import com.ddobang.backend.domain.theme.dto.ThemeStatDto;
 import com.ddobang.backend.domain.theme.entity.Theme;
 
 import lombok.Builder;
 
 /**
- * ThemesResponse
- * 테마 다건 조회 응답에 사용할 dto
+ * ThemeResponse
+ * 테마 상세 조회 응답에 사용할 dto
  * @author 100minha
  */
 @Builder
-public record ThemesResponse(
-	Long id,
+public record ThemeDetailResponse(
 	String name,
 	String storeName,
 	int runtime,
 	String recommendedParticipants,
 	List<String> tags,
-	String thumbnailUrl
+	String thumbnailUrl,
+	float officialDifficulty,
+	ThemeStatDto diaryBasedThemeStat
 ) {
-
-	public static ThemesResponse of(Theme theme) {
-		return ThemesResponse.builder()
-			.id(theme.getId())
+	public static ThemeDetailResponse of(Theme theme, ThemeStatDto themeStatDto) {
+		return ThemeDetailResponse.builder()
 			.name(theme.getName())
 			.storeName(theme.getStore().getName())
 			.runtime(theme.getRuntime())
@@ -33,7 +34,9 @@ public record ThemesResponse(
 			.thumbnailUrl(theme.getThumbnailUrl())
 			.tags(theme.getThemeTagMappings().stream()
 				.map(ttm -> ttm.getThemeTag().getName())
-				.toList())
+				.collect(Collectors.toList()))
+			.officialDifficulty(theme.getOfficialDifficulty())
+			.diaryBasedThemeStat(themeStatDto)
 			.build();
 	}
 }
