+++ src/main/java/com/ddobang/backend/domain/theme/dto/response/ThemeForDiaryResponse.java
@@ -1,4 +1,4 @@
-package com.ddobang.backend.domain.theme.dto;
+package com.ddobang.backend.domain.theme.dto.response;
 
 import com.ddobang.backend.domain.theme.entity.Theme;
