+++ src/main/java/com/ddobang/backend/domain/theme/dto/response/ThemesResponse.java
@@ -1,30 +1,30 @@
-package com.ddobang.backend.domain.theme.dto;
+package com.ddobang.backend.domain.theme.dto.response;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 import com.ddobang.backend.domain.theme.entity.Theme;
 
 import lombok.Builder;
 
 /**
- * ThemeResponse
- * 테마 상세 조회 응답에 사용할 dto
+ * ThemesResponse
+ * 테마 다건 조회 응답에 사용할 dto
  * @author 100minha
  */
 @Builder
-public record ThemeDetailResponse(
+public record ThemesResponse(
+	Long id,
 	String name,
 	String storeName,
 	int runtime,
 	String recommendedParticipants,
 	List<String> tags,
-	String thumbnailUrl,
-	float officialDifficulty,
-	ThemeStatDto diaryBasedThemeStat
+	String thumbnailUrl
 ) {
-	public static ThemeDetailResponse of(Theme theme, ThemeStatDto themeStatDto) {
-		return ThemeDetailResponse.builder()
+
+	public static ThemesResponse of(Theme theme) {
+		return ThemesResponse.builder()
+			.id(theme.getId())
 			.name(theme.getName())
 			.storeName(theme.getStore().getName())
 			.runtime(theme.getRuntime())
@@ -33,9 +33,7 @@ public record ThemeDetailResponse(
 			.thumbnailUrl(theme.getThumbnailUrl())
 			.tags(theme.getThemeTagMappings().stream()
 				.map(ttm -> ttm.getThemeTag().getName())
-				.collect(Collectors.toList()))
-			.officialDifficulty(theme.getOfficialDifficulty())
-			.diaryBasedThemeStat(themeStatDto)
+				.toList())
 			.build();
 	}
 }
