+++ src/main/java/com/ddobang/backend/domain/theme/repository/ThemeRepository.java
@@ -7,14 +7,14 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
-import com.ddobang.backend.domain.theme.dto.ThemeForDiaryResponse;
+import com.ddobang.backend.domain.theme.dto.response.ThemeForDiaryResponse;
 import com.ddobang.backend.domain.theme.entity.Theme;
 
 @Repository
 public interface ThemeRepository extends JpaRepository<Theme, Long>, ThemeRepositoryCustom {
 
 	@Query("""
-		SELECT new com.ddobang.backend.domain.theme.dto.ThemeForDiaryResponse(t.id,t.name,s.name)
+		SELECT new com.ddobang.backend.domain.theme.dto.response.ThemeForDiaryResponse(t.id,t.name,s.name)
 		FROM Theme t JOIN t.store s
 		WHERE t.status != 'DELETED' AND (
 			LOWER(t.name) LIKE LOWER(CONCAT('%', :keyword, '%')) OR
